Corrigido erros

+++ src/server/game/Quests/QuestDef.cpp
@@ -35,10 +35,10 @@ Quest::Quest(Field* questRecord)
     Level = questRecord[2].Get<int16>();
     MinLevel = questRecord[3].Get<uint8>();
     ZoneOrSort = questRecord[4].Get<int16>();
-    Type = questRecord[5].Get<uint16>();
+    Type = questRecord[5].Get<int32>();
     SuggestedPlayers = questRecord[6].Get<uint8>();
     TimeAllowed = questRecord[7].Get<uint32>();
-    AllowableRaces = questRecord[8].Get<uint16>();
+    AllowableRaces = questRecord[8].Get<int32>();
     RequiredFactionId1 = questRecord[9].Get<uint16>();
     RequiredFactionId2 = questRecord[10].Get<uint16>();
     RequiredFactionValue1 = questRecord[11].Get<int32>();
